+++ src/tooling/piral-cli-webpack5/src/webpack/pilet.ts
@@ -21,8 +21,6 @@ async function getConfig(
   sourceMaps = true,
   contentHash = true,
   minimize = true,
-  publicPath = '/',
-  progress = false,
 ): Promise<DefaultConfiguration> {
   const production = !develop;
   const name = process.env.BUILD_PCKG_NAME;
@@ -54,7 +52,7 @@ async function getConfig(
       },
 
       output: {
-        publicPath,
+        publicPath: './',
         path: dist,
         filename: '[name].js',
         chunkFilename: contentHash ? '[chunkhash:8].js' : undefined,
@@ -87,7 +85,7 @@ async function getConfig(
         ],
       },
 
-      plugins: getPlugins([], progress, production, true),
+      plugins: getPlugins([], production, true),
     },
     enhance,
   ];